Documenting Transmission: The Rake Cycle
This is the TEI document used to encode the audio and lyrics to thirteen songs from the 'Rake Cycle' of folk songs as part of a Digital Humanities Master's Thesis. These songs are encoded according to the P5 TEI guidelines. This document was displayed using The Versioning Machine (v-machine.org), an interface for the close comparison and reading of multiple texts.